Fix logic
My dearest,
It's been a long day, and a hard day's work. I just finished working (hopefully for the last time) on the logic for the cargo positions. Please give my love to my children, because it looks that I will be here a long time.
P.S. tell my wife that I shat in most of my pants.